Don’t do anything if there is no selection region
On some occasions (on startup and on executing some commands such as `lsp_document_symbols` which in turn executes `lsp_selection_clear`), the `show_unicode` function might be executed with an empty selection set (i.e. not even one empty region). Previously the function would try to call `view.sel()[0]`, which would generate an IndexError in that case.